+++ src/components/utilities/Portal.tsx
@@ -4,15 +4,16 @@
  * SPDX-License-Identifier: AGPL-3.0-only
  */
 
-import React from 'react';
+import React, { useContext } from 'react';
 import ReactDOM from 'react-dom';
+import { ThemeContext } from '../../theme/theme-context-provider';
 
 interface PortalProps {
 	/** The children to render into the `container` */
 	children: React.ReactElement;
 	/**
 	 * HTML node where to insert the Portal's children.
-	 * The default value is 'window.top.document'.
+	 * The default value is 'windowObj.document'.
 	 * */
 	container?: Element;
 	/** Flag to show or hide Portal's content */
@@ -22,19 +23,16 @@ interface PortalProps {
 }
 
 const Portal = React.forwardRef<React.ReactPortal, PortalProps>(function PortalFn(
-	{
-		children,
-		container = window.top?.document.body || document.body,
-		show = false,
-		disablePortal = false
-	},
+	{ children, container, show = false, disablePortal = false },
 	ref
 ): React.ReactElement | null {
+	const { windowObj } = useContext(ThemeContext);
+
 	if (!show) return null;
 
 	if (disablePortal) return children;
 
-	return ReactDOM.createPortal(children, container);
+	return ReactDOM.createPortal(children, container || windowObj.document.body);
 });
 
 export { Portal, PortalProps };